LayoutTests/fast/forms/select/listbox-click-on-scrollbar-expected.txt

Click on a listbox scrollbar should not dispatch change evnet.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Click on the scrollbar. No change event should be fired.
PASS selectElement.selectedIndex is -1
PASS changeEventCounter is 0

Click on the first option. A change event should be fired.
A change event was dispatched.
PASS selectElement.selectedIndex is 0
PASS changeEventCounter is 1

Add an option item and then click on the scrollbar again. No change event should be fired.
PASS selectElement.selectedIndex is 0
PASS changeEventCounter is 1

PASS successfullyParsed is true

TEST COMPLETE
